requirements.txt
fastapi==0.115.9
uvicorn==0.29.0
streamlit==1.34.0
tensorflow==2.19.0
numpy==1.26.4
Pillow==10.3.0
requests==2.31.0
